fix: fix a name collision in matches

+++ CoqOfRust/examples/axiomatized/examples/ink_contracts/erc721.v
@@ -899,7 +899,7 @@ Section Impl_erc721_Erc721_t.
   
           let count = owned_tokens_count
               .get(from)
-              .map(|c| c - 1)
+              .map(|c| c - (1 as u32))
               .ok_or(Error::CannotFetchValue)?;
           owned_tokens_count.insert( *from, count);
           token_owner.remove(id);
